Update sidebar routes and rename events to box
