+++ src/app/(app)/box/[id]/page.tsx
@@ -28,9 +28,9 @@ export default async function Event({ params }: { params: { id: string } }) {
   return (
     <>
       <div className="max-lg:hidden">
-        <Link href="/events" className="inline-flex items-center gap-2 text-sm/6 text-zinc-500 dark:text-zinc-400">
+        <Link href="/box" className="inline-flex items-center gap-2 text-sm/6 text-zinc-500 dark:text-zinc-400">
           <ChevronLeftIcon className="size-4 fill-zinc-400 dark:fill-zinc-500" />
-          Events
+          Box
         </Link>
       </div>
       <div className="mt-4 flex flex-wrap items-end justify-between gap-4">
